ingestion/requirements.txt
google-api-python-client==2.129.0
yt-dlp==2025.1.26
requests==2.32.3
python-dateutil==2.9.0.post0
dateparser==1.2.0
rapidfuzz==3.9.6
# Use a Python 3.13-compatible psycopg binary build
psycopg[binary]>=3.2.2,<3.3
SQLAlchemy==2.0.35
pydantic==2.8.2
loguru==0.7.2
tenacity==9.0.0
geojson==3.1.0
python-dotenv>=1.0.1,<2.0
